requirements.txt
absl-py==0.8.1
asgiref==3.2.3
astor==0.8.0
attrs==19.3.0
blis==0.4.1
catalogue==1.0.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cmake==3.13.2
confuse==1.0.0
cycler==0.10.0
cymem==2.0.3
decorator==4.4.1
Django==3.0.2
dlib==19.18.0
easydict==1.9
Flask==1.1.1
Flask-Cors==3.0.8
gast==0.2.2
google-pasta==0.1.7
graphviz==0.8.4
grpcio==1.24.1
gunicorn==19.9.0
h5py==2.10.0
idna==2.8
importlib-metadata==0.23
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.0
json5==0.8.5
jsonschema==3.1.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mongoengine==0.18.2
more-itertools==7.2.0
murmurhash==1.0.2
mxnet==1.5.1.post0
networkx==2.4
numpy==1.18.0
opencv-contrib-python==4.1.2.30
opencv-python==4.1.2.30
opt-einsum==3.1.0
pandas==0.25.1
Pillow==6.2.1
plac==1.1.3
preshed==3.0.2
protobuf==3.10.0
pymongo==3.9.0
pyparsing==2.4.6
pyrsistent==0.15.4
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
scikit-learn==0.22.1
scipy==1.3.1
seaborn==0.9.0
Shapely==1.6.4.post2
six==1.13.0
sklearn==0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
spacy==2.2.3
sqlparse==0.3.0
srsly==1.0.1
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
thinc==7.3.1
torch==1.3.1
torchvision==0.4.2
tqdm==4.41.0
urllib3==1.25.6
wasabi==0.6.0
Werkzeug==0.16.0
wrapt==1.11.2
zipp==0.6.0